adjust parameters package structure
