+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/parameter/ShardingParameterRewriterBuilder.java
@@ -20,10 +20,10 @@ package org.apache.shardingsphere.core.rewrite.feature.sharding.parameter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.core.metadata.table.TableMetas;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.parameter.impl.PaginationParameterRewriter;
-import org.apache.shardingsphere.core.rewrite.parameter.ParameterBuilder;
-import org.apache.shardingsphere.core.rewrite.parameter.ParameterRewriter;
-import org.apache.shardingsphere.core.rewrite.parameter.ParameterRewriterBuilder;
-import org.apache.shardingsphere.core.rewrite.parameter.impl.GroupedParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.builder.impl.GroupedParameterBuilder;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.parameter.impl.GeneratedKeyInsertValueParameterRewriter;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.SQLRouteResultAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.ShardingRuleAware;
